POC: IMPALA-7085: Extend CCTZ's interface with get_offset()
Add a function that returns a timezone's UTC offset at given timepoint.
The documentation of CCTZ explicitly warns against doing this, but
this seemed to be the simplest way to speed up UTC->local conversion
in Impala.

Impala stores timestamps as date + "nanoseconds from midnight" pairs
(TimestampValue), while CCTZ's UTC to local conversion (BreakTime)
converts a second from epoch to a struct with separate
year,month...minutes,seconds fields (civil_second).

Impala converts this back into TimestampValue after time-zone
conversion, which makes this detailed break up of timestamps a wasted
effort. Returning the offset without creating/shifting civil_seconds
and adding the offset in Impala lead to a 2x improvement in Impala's
utc->local conversion benchmark (tested with CET timezone). The effect
is even greater for fixed offset timezones due known performance issues
in CCTZ.

Another optimization in get_offset is that the caller has to
provide a "hint" for the lookup to avoid binary search if the
subsequent timestamps fall to the the same timezone rule. BreakTime
uses an atomic integer for this, which impacts multi-threaded
performance.

Testing:
- no automatic tests yet - it seems fairly trivial to add fuzz tests
  to Impala to check that the old and the new timezone conversion
  functions return the same result